Add good enough solution parameter. Multiple refactorings
Add good enough solution parameter to stop execution when a solution hasn't changed in a user specified number of iterations(default 100).
Rename multiple variables to be more in convention with JavaScript practices.
Replace function declarations with function expressions. Arrow functions were considered but not used as they are not supported in IE11.
Replace 'var' with 'let' for variables declared within functions.